Get loose app config file
